Refactor: Use dispatcher and base store instead of EventEmitter.
