+++ app/action-identifiers.js
@@ -1,0 +1,8 @@
+var actionIdentifiers = {
+  articleList: {
+    initialize: 'articleList.initialize',
+    filterByIntensity: 'articleList.filterByIntensity'
+  }
+};
+
+module.exports = actionIdentifiers;
