+++ app/main.js
@@ -1,11 +1,11 @@
 var React = require('react');
-var Store = require('./store.js');
+var ArticleStore = require('./article-store.js');
 var ActionCreator = require('./action-creator.js');
 var App = require('./components/app.react.js');
 var DataAccess = require('./data-access.js');
 
 var dataAccess = new DataAccess();
 var actionCreator = new ActionCreator(dataAccess);
-var store = new Store(actionCreator);
+var store = new ArticleStore();
 
 React.render(<App store={store} actionCreator={actionCreator} />, document.body);
