+++ tests/scenarios/filter-articles.test.js
@@ -20,7 +20,7 @@ describe('filter articles', () => {
     };
     var ActionCreator = require('../../app/action-creator.js');
     var actionCreator = new ActionCreator(dataAccess);
-    var Store = require('../../app/store.js');
+    var Store = require('../../app/article-store.js');
     var store = new Store(actionCreator);
 
     var renderedComponent = React.render(<ComponentClass actionCreator={actionCreator} store={store}/>, renderTarget);
